requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0